Splits up test file, adds more comments and cleaning up small things
